refactor: distinguish between ui, domain and event logic
