+++ src/components/mod.rs
@@ -1,5 +1,6 @@
-pub mod component;
-mod style;
+mod component;
+pub mod style;
+pub mod views;
 pub mod widgets;
 
 pub use component::Component;
