+++ src/components/views/mod.rs
@@ -1,0 +1,5 @@
+pub enum View {
+    Splash,
+    Home,
+    Login,
+}
