+++ src/domain/context.rs
@@ -5,28 +5,28 @@ use std::sync::Mutex;
 use tui::backend::{Backend, CrosstermBackend};
 use tui::terminal::Terminal;
 
-use self::route::Route;
+use super::router::Router;
 use crate::event;
 
-pub struct App<B>
+pub struct Context<B>
 where
     B: Backend,
 {
     terminal: RefCell<Terminal<B>>,
-    route: Route,
+    router: Router,
     pub events: Mutex<event::Events>,
 }
 
-impl App<CrosstermBackend<io::Stdout>> {
+impl Context<CrosstermBackend<io::Stdout>> {
     pub fn new() -> Result<Self> {
         let stdout = io::stdout();
         let backend = CrosstermBackend::new(stdout);
         let mut terminal = Terminal::new(backend)?;
         terminal.hide_cursor()?;
 
-        Ok(App {
+        Ok(Context {
             terminal: RefCell::new(terminal),
-            route: Route::default(),
+            router: Router::default(),
             events: Mutex::new(event::Events::default()),
         })
     }
