+++ src/domain/mod.rs
@@ -1,0 +1,4 @@
+mod context;
+mod router;
+
+pub mod app;
